pyproject.toml
[tool.poetry]
name = "fastapi-service"
version = "0.1.0"
description = "A FastAPI service to convert a video frame to black and white."
authors = ["Guillaume Alleon <guillaume.alleon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.0"
uvicorn = "^0.13.0"
opencv-python-headless = "^4.5.1"
Pillow = "^8.2.0"
python-multipart = "^0.0.5"
aiofiles = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

